feat: nowails build tag — chromium-shell binary with zero GTK linkage
Phase 4 of docs/chromium-shell.md. The Wails shell moves behind a build
tag so a Chromium-only binary carries no Wails — and with it no
GTK/WebKitGTK — linkage at all:

- shell_wails.go (!nowails): runWails, the Wails event-bridge fallback
  and the Wails dialog picker choice, verbatim from main.go.
- shell_wails_stub.go (nowails): wailsShellAvailable=false routes main
  to the Chromium shell unconditionally; runWails fatals with a clear
  message if reached.
- project.WailsPicker moves to picker_wails.go behind the same tag;
  ZenityPicker (pure Go, shells out) is always available.
- task build:chromium: CGO_ENABLED=0 go build -tags nowails → a fully
  static bin/lich-chromium (every remaining dep is CGO-free; ldd says
  "not a dynamic executable").

Formal packaging (nfpm/AppImage swap, dropping fix-appimage.sh and the
bundled WebKitGTK) deliberately waits for phase 5 — shipping it now
would mean two release pipelines for one transition window.
