+++ docs/chromium-shell.md
@@ -76,10 +76,20 @@ Migration progress:
    WebKitGTK workarounds were not ported (they patched ghostty-web 0.4.0
    bugs); ghostty itself stays reachable via
    localStorage.setItem("lich.terminal", "ghostty") until phase 5.
-4. **Packaging**: AppImage/deb/rpm ship only the Go binary; drop
-   `fix-appimage.sh` and the bundled WebKitGTK entirely.
+4. **De-Wails the build** — DONE (phase 4): `-tags nowails` produces a binary
+   with zero Wails/GTK/WebKitGTK linkage — pure Go, fully static under
+   `CGO_ENABLED=0` (modernc sqlite, creack/pty, coder/websocket, zenity are
+   all CGO-free). `task build:chromium` builds it (`bin/lich-chromium`,
+   Chromium shell unconditionally; `LICH_SHELL` is ignored — there is nothing
+   else to run). The Wails path lives in `shell_wails.go` / the picker in
+   `picker_wails.go`, both behind `!nowails`. Formal packaging (nfpm/
+   AppImage swap, dropping `fix-appimage.sh` and the bundled WebKitGTK) lands
+   with phase 5 — doing it now would mean maintaining two release pipelines
+   for one transition window.
 5. **Cleanup**: default the shell to Chromium, delete the Wails path, the
-   ghostty patches and the GDK_BACKEND hack.
+   ghostty patches and the GDK_BACKEND hack; release pipeline ships the
+   static binary (AppImage becomes a trivial wrapper, .deb/.rpm/.pkg just
+   carry one file).
 
 What dies with WebKitGTK (all "Known Ceilings" entries): forced
 `GDK_BACKEND=x11`, the sandbox-disabled AppImage, the contenteditable DOM
