+++ internal/project/picker.go
@@ -4,28 +4,7 @@ import (
 	"errors"
 
 	"github.com/ncruces/zenity"
-	"github.com/wailsapp/wails/v3/pkg/application"
 )
-
-// WailsPicker asks through the Wails dialog API — only valid while the Wails
-// application shell is running.
-type WailsPicker struct{}
-
-func (WailsPicker) PickDirectory(title string) (string, error) {
-	return application.Get().Dialog.OpenFile().
-		CanChooseDirectories(true).
-		CanChooseFiles(false).
-		SetTitle(title).
-		PromptForSingleSelection()
-}
-
-func (WailsPicker) PickFile(title string) (string, error) {
-	return application.Get().Dialog.OpenFile().
-		CanChooseFiles(true).
-		CanChooseDirectories(false).
-		SetTitle(title).
-		PromptForSingleSelection()
-}
 
 // ZenityPicker shells out to zenity/qarma (github.com/ncruces/zenity) — the
 // Chromium shell path, where no Wails application exists. Cancel maps to
